Update an old post, thanks to alert reader Marc Singer!
I assume this code used to work even for zero elements in the array,
but I confirm that it does *not* work anymore, in either GCC or Clang.
And no older releases on Godbolt seem to work either. So maybe it
changed, or maybe it never worked, I'm not sure. Anyway, fixed now.