Tag types (Cunningham's Law)
Drive-by: There is still "poison pill" wording in https://eel.is/c++draft/concept.swappable#2.1
(even with a note from 2022 about why it's needed, which I suspect is still wrong),
but I don't recall if that's even the wording I was objecting to in 2019,
and removing it certainly wouldn't make the spec of ranges::swap noticeably
simpler. So there's no need to mention it in the entry for CPOs.